pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "livekit-agents[openai,turn-detector,silero,cartesia,deepgram]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "httpx",
    # Phase 1 – Dashboard
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.21",
    "greenlet>=3.0",
    # Phase 2 – Gmail
    "google-auth>=2.38",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.165",
    "anthropic>=0.49",
    # Phase 3 – Slack
    "slack-bolt>=1.21",
    "slack-sdk>=3.34",
    # Phase 4 – Discord
    "discord.py>=2.4",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
